--- src/client/components/SidebarMenu/DesktopVersion/SidebarMenuDesktop.css
@@ -1,33 +1,0 @@
-.sidebar-menu-wrapper-desktop {
-  display: flex;
-  justify-content: center;
-  left: 0;
-  width: 3.1rem;
-  height: 100vh;
-  background-color: rgba(83, 82, 82, 0.849);
-  position: fixed;
-}
-
-.sidebar-menu-button-desktop {
-  background: none;
-  color: inherit;
-  border: none;
-  cursor: pointer;
-  outline: inherit;
-  display: block;
-  padding: 1rem 0.7rem;
-  overflow: hidden;
-}
-
-.sidebar-menu-button-desktop:hover {
-  background-color: rgb(71, 71, 71);
-}
-
-.sidebar-menu-desktop {
-  top: 20%;
-  position: fixed;
-}
-
-.sidebar-menu-icon-width-desktop {
-  font-size: 1.3rem;
-}
