+++ src/client/components/SidebarMenu/SidebarMenu.story.js
@@ -1,12 +1,5 @@
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import SidebarMenuMobile from './MobileVersion/SidebarMenuMobile.component';
-import SidebarMenuDesktop
-  from './DesktopVersion/SidebarMenuDesktop.component';
-import SidebarMenuResp from './responsive/SidebarMenuResp.component';
+import SidebarMenu from './SidebarMenu.component';
 
-storiesOf ('SidebarMenu', module).add ('mobile version', () => <SidebarMenuMobile />);
-storiesOf ('SidebarMenu', module).add ('desktop version', () => (
-  <SidebarMenuDesktop />
-));
-storiesOf ('SidebarMenu', module).add('responsive', () => <SidebarMenuResp />)
+storiesOf ('SidebarMenu', module).add('responsive', () => <SidebarMenu />)
